Rename ArgumentTest to ArgumentTypeTest and add some actual argument tests
