+++ src/main/java/net/minestom/server/listener/TabCompleteListener.java
@@ -30,7 +30,7 @@ public class TabCompleteListener {
                     suggestion.getStart(),
                     suggestion.getLength(),
                     suggestion.getEntries().stream()
-                            .map(suggestionEntry -> new TabCompletePacket.Match(suggestionEntry.getEntry(), suggestionEntry.getTooltip()))
+                            .map(suggestionEntry -> new TabCompletePacket.Match(suggestionEntry.entry(), suggestionEntry.tooltip()))
                             .toList())
             );
         }
